Add GridViewRowPresenter attempt
The initial idea was to do a code-behind hack, which mostly worked,
but it turns out the "bind to the other column widths" trick from
the first attempt works here too.  I think this actually works a bit
better than the other approach.

Also, made the faux highlighting in the previous thing set the
background on the entire row, not just the last cell.