Improve generification
Replaced the ListView element name references with a more general
RelativeSource binding.  Moved the RowPresenter variant to a style
resource.  The idea is to be able to place the style in a
separate file, although it's of dubious genericity because the
column in which the long item starts is hard-coded.

Also, #ifdefed out the column width tracking stuff.